+++ tests/Benchmark/ObjectDataSetBench.php
@@ -5,24 +5,22 @@ declare(strict_types=1);
 namespace Yiisoft\Validator\Tests\Benchmark;
 
 use Generator;
+use PhpBench\Benchmark\Metadata\Annotations\BeforeMethods;
 use PhpBench\Benchmark\Metadata\Annotations\Iterations;
 use PhpBench\Benchmark\Metadata\Annotations\ParamProviders;
 use PhpBench\Benchmark\Metadata\Annotations\Revs;
-use Yiisoft\Validator\DataSet\CacheObjectDataSetDecorator;
+use ReflectionProperty;
 use Yiisoft\Validator\DataSet\ObjectDataSet;
 use Yiisoft\Validator\Tests\Stub\ObjectWithDifferentPropertyVisibility;
 
 class ObjectDataSetBench
 {
+    private static $currentPropertyVisibility = ReflectionProperty::IS_PUBLIC;
+
     public function provideObjectDataSets(): Generator
     {
-        yield 'compare without cache' => [
+        yield 'one instance' => [
             'dataSet' => new ObjectDataSet(new ObjectWithDifferentPropertyVisibility()),
-        ];
-        yield 'compare with cache' => [
-            'dataSet' => new CacheObjectDataSetDecorator(
-                new ObjectDataSet(new ObjectWithDifferentPropertyVisibility())
-            ),
         ];
     }
 
@@ -41,7 +39,7 @@ class ObjectDataSetBench
      * @Revs(100000)
      * @Iterations(5)
      */
-    public function benchGetRulesWithNewInstanceAndNoCache(): void
+    public function benchGetRulesWithNewInstanceAndCache(): void
     {
         $dataSet = new ObjectDataSet(new ObjectWithDifferentPropertyVisibility());
         $dataSet->getRules();
@@ -51,10 +49,15 @@ class ObjectDataSetBench
      * @Revs(100000)
      * @Iterations(5)
      */
-    public function benchGetRulesWithNewInstanceAndCache(): void
+    public function benchGetRulesWithNewInstanceAndNoCache(): void
     {
-        $dataSet = new ObjectDataSet(new ObjectWithDifferentPropertyVisibility());
-        $dataSet = new CacheObjectDataSetDecorator($dataSet);
+        // Changing property visibility allows to clear cache without additional mocks, etc.
+        $propertyVisibility = self::$currentPropertyVisibility === ReflectionProperty::IS_PUBLIC
+            ? ReflectionProperty::IS_PRIVATE
+            : ReflectionProperty::IS_PUBLIC;
+        self::$currentPropertyVisibility = $propertyVisibility;
+
+        $dataSet = new ObjectDataSet(new ObjectWithDifferentPropertyVisibility(), $propertyVisibility);
         $dataSet->getRules();
     }
 }
